update for change in the error message returned for too-large query results, plus test
