+++ tests/integration/integration_test.py
@@ -139,46 +139,6 @@ def test_createEntity_with_provenance():
     # Verify the Provenance
     activity = syn.getProvenance(entity)
     assert activity['used'][0]['reference']['targetId'] == 'syn123'
-
-
-def test_query():
-    ## TODO: replace this test with the one below when query() is replaced
-    
-    # Remove all the Entities that are in the project
-    qry = syn.query("select id from entity where entity.parentId=='%s'" % project['id'])
-    for res in qry['results']:
-        syn.delete(res['entity.id'])
-    
-    # Add entities and verify that I can find them with a query
-    for i in range(2):
-        syn.store(Data(parent=project['id']))
-        qry = syn.query("select id, name from entity where entity.parentId=='%s'" % project['id'])
-        assert qry['totalNumberOfResults'] == i + 1
-
-        
-def test_chunked_query():
-    oldLimit = client.QUERY_LIMIT
-    try:
-        client.QUERY_LIMIT = 3
-        
-        # Remove all the Entities that are in the project
-        iter = syn.chunkedQuery("select id from entity where entity.parentId=='%s'" % project['id'])
-        for res in iter:
-            syn.delete(res['entity.id'])
-        
-        # Dump a bunch of Entities into the project
-        for i in range(client.QUERY_LIMIT * 5):
-            syn.store(Data(parent=project['id']))
-                
-        # Give a bunch of limits/offsets to be ignored (except for the first ones)
-        queryString = "select * from entity where entity.parentId=='%s' offset  1 limit 9999999999999    offset 2345   limit 6789 offset 3456    limit 5689" % project['id']
-        iter = syn.chunkedQuery(queryString)
-        count = 0
-        for res in iter:
-            count += 1
-        assert count == (client.QUERY_LIMIT * 5)
-    finally:
-        client.QUERY_LIMIT = oldLimit
     
 
 def test_md5_query():
